feat: add GH workflow to generate translations
This GH workflow generates translation files from a matrix of
repositories in the openedx organization. After generating the
translation files, it commits them to a branch, and creates a pull
request to be merged. In the case that nothing has been committed, the
branch is deleted, and the user notified.