feat: add GH workflow to generate translations (#5)
This GH workflow generates translation files from a matrix of
repositories in the openedx organization. After generating the
translation files, it commits them to a branch, and creates a pull
request to be merged. In the case that nothing has been committed, the
branch is deleted, and the user notified.

Co-authored-by: Brain Mesick <bmesick@tcril.org>
Co-authored-by: Feanil Patel <feanil@tcril.org>