chore: update how env vars and outputs are used (#11)
* chore: update how env vars and outputs are used

For more information see: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#environment-files

* feat: add wc -l to git status output

The env variables did not accept the raw output of git status. By adding
wc -l, it counts the number of lines. And since we are only asking for
the git status of two files, as long as it is greater than 0, it is
sufficient to commit.