fix: remove always delete branch (#86)
Always deleting the branch whether it succeeds or fails was causing the
PR to not merge even if it was successful because the branch was
deleted. Automatic branch deletion once it has been merged has been
added to the repo's settings, and the deletion step has been absorbed by
the previous notification step so that it only deletes branches without
commits.