fix: retry merge PR step for transifex automerge (#222)
When multiple PRs are created by the transifex github app integration at once, it is common for one or more to fail due to an out of sync base branch. This updates the automerge action to verify the PR has been merged using `gh pr status`, and retries if the PR has not been merged.